refactor: extract CopyButton component
