+++ renderer/components/UI/index.js
@@ -13,6 +13,7 @@ export Button from './Button'
 export ButtonCreate from './ButtonCreate'
 export ClippedText from './ClippedText'
 export CopyBox from './CopyBox'
+export CopyButton from './CopyButton'
 export SwitchButton from './SwitchButton'
 export Card from './Card'
 export Checkbox from './Checkbox'
